+++ backend/epic_app/serializers/__init__.py
@@ -1,7 +1,10 @@
 # Expose all neede serializers here.
 from epic_app.serializers.agency_serializer import AgencySerializer
 from epic_app.serializers.area_serializer import AreaSerializer
-from epic_app.serializers.epic_user_serializer import EpicUserSerializer
+from epic_app.serializers.epic_user_serializer import (
+    EpicOrganizationSerializer,
+    EpicUserSerializer,
+)
 from epic_app.serializers.group_serializer import GroupSerializer
 from epic_app.serializers.program_serializer import ProgramSerializer
 from epic_app.serializers.question_serializer import (