+++ backend/epic_app/serializers/answer_serializer.py
@@ -5,17 +5,23 @@ from epic_app.models.epic_answers import (
     MultipleChoiceAnswer,
     SingleChoiceAnswer,
     YesNoAnswer,
+    YesNoAnswerType,
 )
+from epic_app.models.epic_questions import EvolutionChoiceType
 from epic_app.models.epic_user import EpicUser
 
 
 class YesNoAnswerSerializer(serializers.ModelSerializer):
+    short_answer = serializers.ChoiceField(YesNoAnswerType.choices)
+
     class Meta:
         model = YesNoAnswer
         fields = ("url", "id", "user", "question", "short_answer", "justify_answer")
 
 
 class SingleChoiceAnswerSerializer(serializers.ModelSerializer):
+    selected_choice = serializers.ChoiceField(EvolutionChoiceType.choices)
+
     class Meta:
         model = SingleChoiceAnswer
         fields = ("url", "id", "user", "question", "selected_choice", "justify_answer")