+++ backend/epic_app/tests/serializers/test_answer_serializer.py
@@ -15,7 +15,7 @@ from epic_app.models.epic_questions import (
     LinkagesQuestion,
     NationalFrameworkQuestion,
 )
-from epic_app.models.epic_user import EpicUser
+from epic_app.models.epic_user import EpicOrganization, EpicUser
 from epic_app.models.models import Program
 from epic_app.serializers.answer_serializer import (
     MultipleChoiceAnswerSerializer,
@@ -37,7 +37,8 @@ def answer_serializer_fixture(
         epic_test_db (pytest.fixture): Fixture to load for the whole file tests.
     """
     theonewhoasks = EpicUser.objects.create(
-        username="TheOneWhoAsks", organization="TestCorp"
+        username="TheOneWhoAsks",
+        organization=EpicOrganization.objects.create(name="TestCorp"),
     )
     YesNoAnswer.objects.create(
         user=theonewhoasks,
@@ -83,7 +84,7 @@ class TestAnswerSerializer:
                     "id": 1,
                     "user": 4,
                     "question": 1,
-                    "short_answer": "Y",
+                    "short_answer": str(YesNoAnswerType.YES),
                     "justify_answer": "Velit ex cupidatat do magna ipsum.",
                 },
             ),
@@ -94,7 +95,7 @@ class TestAnswerSerializer:
                     "id": 2,
                     "justify_answer": "Ipsum anim fugiat sit nostrud enim.",
                     "question": 3,
-                    "selected_choice": "Engaged",
+                    "selected_choice": str(EvolutionChoiceType.ENGAGED),
                     "url": "http://testserver/api/singleanswer/2/",
                     "user": 4,
                 },